refactor: Entity에 사용되는 인덱스 정의 및 클래스명 수정
- 테스트 이름 수정
