+++ src/test/java/com/nexters/phochak/service/LikesServiceTest.java
@@ -26,7 +26,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class LikesServiceImplTest {
+class LikesServiceTest {
 
     @InjectMocks
     LikeServiceImpl likeService;
